Split UserAccountService into UserAccountService and TokenService
The token part doesn't need a DB access which makes UserAccountService
not easily to use when only needing the token part.
