+++ src/service/mod.rs
@@ -9,6 +9,7 @@ pub mod invoice_line;
 pub mod item;
 pub mod name;
 pub mod stock_line;
+pub mod token;
 pub mod token_bucket;
 pub mod user_account;
